added admin

+++ server/modules/api/categories/categoryController.js
@@ -23,34 +23,49 @@ Router.get("/:name", (req, res) => {
   });
 });
 
-Router.post("/", (req, res) => {
-  categoryModel.createNewCategory(req.body.name, req.body.alias, req.body.keywords, (err, result) => {
-    if (err) {
-      res.send(err);
-    } else {
-      res.send(result);
-    }
-  });
+Router.post("/", passport.authenticate('jwt', { session: false }), (req, res) => {
+  if (req.user.admin) {
+    categoryModel.createNewCategory(req.body.name, req.body.alias, req.body.keywords, (err, result) => {
+      if (err) {
+        res.send(err);
+      } else {
+        res.send(result);
+      }
+    });
+  } else {
+    res.status(400);
+    res.send("Unauthorized");
+  }
 });
 
-Router.put("/:id", (req, res) => {
-  categoryModel.updateCategoryById(req.params.id, req.body, (err, result) => {
-    if (err) {
-      res.send(err);
-    } else {
-      res.send(result);
-    }
-  });
+Router.put("/:id", passport.authenticate('jwt', { session: false }), (req, res) => {
+  if (req.user.admin) {
+    categoryModel.updateCategoryById(req.params.id, req.body, (err, result) => {
+      if (err) {
+        res.send(err);
+      } else {
+        res.send(result);
+      }
+    });
+  } else {
+    res.status(400);
+    res.send("Unauthorized");
+  }
 });
 
-Router.delete("/:id", (req, res) => {
-  categoryModel.deleteCategoryById(req.params.id, (err, result) => {
-    if (err) {
-      res.send(err);
-    } else {
-      res.send(result);
-    }
-  });
+Router.delete("/:id", passport.authenticate('jwt', { session: false }), (req, res) => {
+  if (req.user.admin) {
+    categoryModel.deleteCategoryById(req.params.id, (err, result) => {
+      if (err) {
+        res.send(err);
+      } else {
+        res.send(result);
+      }
+    });
+  } else {
+    res.status(400);
+    res.send("Unauthorized");
+  }
 });
 
 module.exports = Router;
