+++ server/modules/api/pages/pageController.js
@@ -13,38 +13,49 @@ Router.get("/all", (req, res) => {
   });
 });
 
-Router.post("/", (req, res) => {
-  console.log("req.body:\n", req.body);
-  console.log("req.body.permalink_url:\n", req.body.permalink_url);
-  console.log("req.body.name:\n", req.body.name);
-  console.log("req.body.category:\n", req.body.category);
-  pageModel.createNewPageWithCategoryId(req.body.permalink_url, req.body.name, req.body.category, (err, result) => {
-    if (err) {
-      res.send(err);
-    } else {
-      res.send(result);
-    }
-  });
+Router.post("/", passport.authenticate('jwt', { session: false }), (req, res) => {
+  if (req.user.admin) {
+    pageModel.createNewPageWithCategoryId(req.body.permalink_url, req.body.name, req.body.category, (err, result) => {
+      if (err) {
+        res.send(err);
+      } else {
+        res.send(result);
+      }
+    });
+  } else {
+    res.status(400);
+    res.send("Unauthorized");
+  }
 });
 
-Router.put("/:id", (req, res) => {
-  pageModel.updatePageById(req.params.id, req.body, (err, result) => {
-    if (err) {
-      res.send(err);
-    } else {
-      res.send(result);
-    }
-  });
+Router.put("/:id", passport.authenticate('jwt', { session: false }), (req, res) => {
+  if (req.user.admin) {
+    pageModel.updatePageById(req.params.id, req.body, (err, result) => {
+      if (err) {
+        res.send(err);
+      } else {
+        res.send(result);
+      }
+    });
+  } else {
+    res.status(400);
+    res.send("Unauthorized");
+  }
 });
 
-Router.delete("/:id", (req, res) => {
-  pageModel.deletePageById(req.params.id, (err, result) => {
-    if (err) {
-      res.send(err);
-    } else {
-      res.send(result);
-    }
-  });
+Router.delete("/:id", passport.authenticate('jwt', { session: false }), (req, res) => {
+  if (req.user.admin) {
+    pageModel.deletePageById(req.params.id, (err, result) => {
+      if (err) {
+        res.send(err);
+      } else {
+        res.send(result);
+      }
+    });
+  } else {
+    res.status(400);
+    res.send("Unauthorized");
+  }
 });
 
 Router.get("/:url", (req, res) => {
